+++ source/Messaging.Tests/Application/OutgoingMessages/RejectRequestChangeOfSupplier/RejectRequestChangeOfSupplierDocumentWriterTests.cs
@@ -16,82 +16,84 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Net.Mime;
 using System.Threading.Tasks;
 using System.Xml.Linq;
 using System.Xml.Schema;
 using Messaging.Application.Common;
 using Messaging.Application.Configuration;
-using Messaging.Application.OutgoingMessages.AccountingPointCharacteristics;
+using Messaging.Application.OutgoingMessages.RejectRequestChangeOfSupplier;
 using Messaging.Application.SchemaStore;
 using Messaging.Domain.OutgoingMessages;
 using Messaging.Infrastructure.Common;
 using Messaging.Infrastructure.Configuration;
 using Messaging.Infrastructure.Configuration.Serialization;
-using Messaging.Tests.OutgoingMessages.Asserts;
+using Messaging.Tests.Application.OutgoingMessages.Asserts;
 using Xunit;
-using MarketActivityRecord = Messaging.Application.OutgoingMessages.AccountingPointCharacteristics.MarketActivityRecord;
 
-namespace Messaging.Tests.OutgoingMessages.AccountingPointCharacteristics;
+namespace Messaging.Tests.Application.OutgoingMessages.RejectRequestChangeOfSupplier;
 
-public class AccountingPointCharacteristicsDocumentWriterTests
+public class RejectRequestChangeOfSupplierDocumentWriterTests
 {
-    private readonly AccountingPointCharacteristicsDocumentWriter _documentWriter;
+    private readonly RejectRequestChangeOfSupplierDocumentWriter _documentWriter;
     private readonly ISystemDateTimeProvider _systemDateTimeProvider;
     private readonly IMarketActivityRecordParser _marketActivityRecordParser;
-    private readonly SampleData _sampleData;
     private ISchemaProvider? _schemaProvider;
 
-    public AccountingPointCharacteristicsDocumentWriterTests()
+    public RejectRequestChangeOfSupplierDocumentWriterTests()
     {
         _systemDateTimeProvider = new SystemDateTimeProvider();
         _marketActivityRecordParser = new MarketActivityRecordParser(new Serializer());
-        _documentWriter = new AccountingPointCharacteristicsDocumentWriter(_marketActivityRecordParser);
-        _sampleData = new SampleData();
+        _documentWriter = new RejectRequestChangeOfSupplierDocumentWriter(_marketActivityRecordParser);
     }
 
     [Fact]
     public async Task Document_is_valid()
     {
-        var header = new MessageHeader("E03", "SenderId", "DDZ", "ReceiverId", "DDQ", Guid.NewGuid().ToString(), _systemDateTimeProvider.Now(), null);
-        var marketActivityRecord = _sampleData.CreateMarketActivityRecord();
+        var header = new MessageHeader("E03", "SenderId", "DDZ", "ReceiverId", "DDQ", Guid.NewGuid().ToString(), _systemDateTimeProvider.Now(), "A01");
         var marketActivityRecords = new List<MarketActivityRecord>()
         {
-            marketActivityRecord,
+            new(Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), "FakeMarketEvaluationPointId", new List<Reason>()
+            {
+                new Reason("Reason1", "999"),
+                new Reason("Reason2", "999"),
+            }),
+            new(Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), "FakeMarketEvaluationPointId",
+            new List<Reason>()
+            {
+                new Reason("Reason3", "999"),
+                new Reason("Reason4", "999"),
+            }),
         };
+
         var message = await _documentWriter.WriteAsync(header, marketActivityRecords.Select(record => _marketActivityRecordParser.From(record)).ToList()).ConfigureAwait(false);
+
         await AssertMessage(message, header, marketActivityRecords).ConfigureAwait(false);
     }
 
     private static void AssertMarketActivityRecords(List<MarketActivityRecord> marketActivityRecords, XDocument document)
     {
-        AssertXmlMessage.AssertMarketActivityRecordCount(document, 1);
+        AssertXmlMessage.AssertMarketActivityRecordCount(document, 2);
         foreach (var activityRecord in marketActivityRecords)
         {
             var marketActivityRecord = AssertXmlMessage.GetMarketActivityRecordById(document, activityRecord.Id)!;
 
             Assert.NotNull(marketActivityRecord);
-          //  AssertXmlMessage.AssertMarketActivityRecordValue(marketActivityRecord, "originalTransactionIDReference_MktActivityRecord.mRID", activityRecord.OriginalTransactionId);
-            AssertXmlMessage.AssertMarketActivityRecordValue(marketActivityRecord, "mRID", activityRecord.Id);
-            AssertXmlMessage.AssertMarketActivityRecordValue(marketActivityRecord, "validityStart_DateAndOrTime.dateTime", activityRecord.ValidityStartDate.ToString());
+            AssertXmlMessage.AssertMarketActivityRecordValue(marketActivityRecord, "originalTransactionIDReference_MktActivityRecord.mRID", activityRecord.OriginalTransactionId);
+            AssertXmlMessage.AssertMarketActivityRecordValue(marketActivityRecord, "marketEvaluationPoint.mRID", activityRecord.MarketEvaluationPointId);
+            AssertXmlMessage.AssertReasons(marketActivityRecord, activityRecord.Reasons.ToList());
         }
     }
 
     private async Task AssertMessage(Stream message, MessageHeader header, List<MarketActivityRecord> marketActivityRecords)
     {
+        _schemaProvider = new XmlSchemaProvider();
         var document = XDocument.Load(message);
         AssertXmlMessage.AssertHeader(header, document);
-        AssertXmlMessage.AssertHasHeaderValue(document, "type", "E07");
+        AssertXmlMessage.AssertHasHeaderValue(document, "type", "E44");
 
         AssertMarketActivityRecords(marketActivityRecords, document);
 
-        await AssertConformsToSchema(message).ConfigureAwait(false);
-    }
-
-    private async Task AssertConformsToSchema(Stream message)
-    {
-        _schemaProvider = new XmlSchemaProvider();
-        var schema = await _schemaProvider.GetSchemaAsync<XmlSchema>("accountingpointcharacteristics", "0.1")
+        var schema = await _schemaProvider.GetSchemaAsync<XmlSchema>("rejectrequestchangeofsupplier", "0.1")
             .ConfigureAwait(false);
         await AssertXmlMessage.AssertConformsToSchemaAsync(message, schema!).ConfigureAwait(false);
     }
